% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gwas_glm.R
\name{run_gwas_glm}
\alias{run_gwas_glm}
\title{Run GWAS with GLM and Enhanced Visualization}
\usage{
run_gwas_glm(y, X, C, snp_info)
}
\arguments{
\item{y}{Phenotype data frame}

\item{X}{Genotype data frame}

\item{C}{Covariate data frame}

\item{snp_info}{SNP metadata data frame}
}
\value{
List containing results and plots
}
\description{
Performs GWAS using linear regression with improved visualization
}
